fix: deprecated and merged am0016 and am0017
